# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "ATP-TDM"
  },
  "security" : [
    {
      "Bearer Authentication" : [ ]
    }
  ],
...
  "paths" : {
...
    "/api/tdm/occupy" : {
      "put" : {
        "description" : "Marks the rows as occupied by the user in occupiedBy and records the occupation in the statistics.",
        "operationId" : "occupyTestData",
        "parameters" : [
          {
            "description" : "Database table name of the test data table.",
            "in" : "query",
            "name" : "tableName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Name of the user who occupies the rows.",
            "in" : "query",
            "name" : "occupiedBy",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "description" : "ROW_ID values of the rows.",
                "items" : {
                  "type" : "string",
                  "format" : "uuid"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Occupy rows by ID",
        "tags" : [
          "test-data-controller"
        ]
      }
    },
...
    "/api/tdm/release" : {
      "put" : {
        "description" : "Makes the rows available again and removes their occupation records from the statistics.",
        "operationId" : "releaseTestData",
        "parameters" : [
          {
            "description" : "Database table name of the test data table.",
            "in" : "query",
            "name" : "tableName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "description" : "ROW_ID values of the rows.",
                "items" : {
                  "type" : "string",
                  "format" : "uuid"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Release rows by ID",
        "tags" : [
          "test-data-controller"
        ]
      }
    },
...
    "/api/tdm/rest/add-info-to-row" : {
      "post" : {
        "description" : "For each request in add-info-to-row-requests, appends each value in record-with-data-for-update on a new line to the current value of its column, in every row that matches search-row-parameters-set, and returns the number of updated rows.",
        "operationId" : "addInfoToRow",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Add values to rows",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
...
    "/api/tdm/rest/get-record" : {
      "post" : {
        "description" : "For each request in get-row-requests, returns the value of name-column-response from the first available row that matches search-row-parameters-set. The row is not occupied.",
        "operationId" : "atpGetTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Read one column of a row",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/get-records" : {
      "post" : {
        "description" : "For each request in get-row-requests, returns the values of response-column-names from the first available row that matches search-row-parameters-set, as a JSON object. The row is not occupied.",
        "operationId" : "getMultipleColumnTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Read several columns of a row",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/insert-records" : {
      "post" : {
        "description" : "Inserts the rows from insert-records. When no table with this title exists, creates the table first.",
        "operationId" : "insertTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Insert rows into a table",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/occupy-records" : {
      "post" : {
        "description" : "For each request in occupy-row-requests, occupies the first available row that matches search-row-parameters-set and returns the value of name-column-response. The row is marked as occupied by ATP_User.",
        "operationId" : "atpOccupyTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Occupy rows and return one column",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/occupy-records-full-row" : {
      "post" : {
        "description" : "For each request in occupy-full-row-requests, occupies the first available row that matches search-row-parameters-set and returns the values of response-column-names as a JSON object. The row stays available when one of the columns does not exist. The row is marked as occupied by ATP_User.",
        "operationId" : "occupyTestDataFullRow",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Occupy rows and return several columns",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/refresh-tables" : {
      "post" : {
        "description" : "Runs the refresh query of the table with this title in the system, or of every table with this title in the project when envName or systemName is missing.",
        "operationId" : "refreshTables",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Refresh tables by title",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/release-records" : {
      "post" : {
        "description" : "For each request in release-row-requests, releases the occupied row that matches search-row-parameters-set and returns the value of name-column-response. The request fails when more than one occupied row matches.",
        "operationId" : "atpReleaseTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Release occupied rows",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/release-records/bulk" : {
      "post" : {
        "operationId" : "releaseFullTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Release all occupied rows of a table",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/resolve-table" : {
      "post" : {
        "description" : "Returns the name of the database table that stores the table with this title. All four fields are required; a missing field or an unknown table returns HTTP 400.",
        "operationId" : "resolveTableName",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get the database table name of a table",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/run-cleanup-table" : {
      "post" : {
        "description" : "Runs the cleanup configured for the table, and returns the number of removed rows. Unlike the other operations, projectName must hold the project ID.",
        "operationId" : "runCleanupForTable",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Run the cleanup of a table",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/truncate-table" : {
      "post" : {
        "description" : "Unlike the other operations, projectName must hold the project ID.",
        "operationId" : "truncateTable",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Delete all rows of a table",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
    "/api/tdm/rest/update-records" : {
      "post" : {
        "description" : "For each request in update-row-requests, sets the columns in record-with-data-for-update to the given values in every row that matches search-row-parameters-set, and returns the number of updated rows.",
        "operationId" : "updateTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestApiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResponseMessage"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Update rows",
        "tags" : [
          "atp-action-controller"
        ]
      }
    },
...
    "/api/tdm/table" : {
      "delete" : {
        "description" : "Drops the database table, removes it from the catalog, and stops its scheduled refresh.",
        "operationId" : "deleteTestData",
        "parameters" : [
          {
            "description" : "Database table name of the test data table.",
            "in" : "query",
            "name" : "tableName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Delete a table",
        "tags" : [
          "test-data-controller"
        ]
      },
      "post" : {
        "description" : "Returns a page of rows that match the filters in the request body, sorted as requested, from the available rows or, with occupied set, from the occupied rows.",
        "operationId" : "getTestData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TestDataRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TestDataTable"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Read rows of a table",
        "tags" : [
          "test-data-controller"
        ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AddInfoToRowRequest" : {
        "type" : "object",
        "properties" : {
          "record-with-data-for-update" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "search-row-parameters-set" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TestDataTableFilter"
            }
          }
        }
      },
...
      "OccupyRowRequest" : {
        "type" : "object",
        "properties" : {
          "name-column-response" : {
            "type" : "string"
          },
          "search-row-parameters-set" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TestDataTableFilter"
            }
          }
        }
      },
...
      "ResponseMessage" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "contentObject" : {
            "type" : "object"
          },
          "link" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [
              "success",
              "error"
            ]
          }
        }
      },
      "RestApiRequest" : {
        "type" : "object",
        "properties" : {
          "add-info-to-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AddInfoToRowRequest"
            }
          },
          "envName" : {
            "type" : "string"
          },
          "environment" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            },
            "writeOnly" : true
          },
          "get-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GetRowRequest"
            }
          },
          "insert-records" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "occupy-full-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OccupyFullRowRequest"
            }
          },
          "occupy-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OccupyRowRequest"
            }
          },
          "projectName" : {
            "type" : "string"
          },
          "release-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReleaseRowRequest"
            }
          },
          "systemName" : {
            "type" : "string"
          },
          "title-table" : {
            "type" : "string"
          },
          "update-row-requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UpdateRowRequest"
            }
          }
        }
      },
...
      "TestDataRequest" : {
        "type" : "object",
        "properties" : {
          "dataTableOrder" : {
            "$ref" : "#/components/schemas/TestDataTableOrder"
          },
          "filters" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TestDataTableFilter"
            }
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "occupied" : {
            "type" : "boolean"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tableName" : {
            "type" : "string"
          }
        }
      },
      "TestDataTable" : {
        "type" : "object",
        "properties" : {
          "columns" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TestDataTableColumn"
            }
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "name" : {
            "type" : "string"
          },
          "query" : {
            "type" : "string"
          },
          "records" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [
              "ALL",
              "AVAILABLE",
              "OCCUPIED"
            ]
          },
          "updateByQuery" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}